+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove use of deprecated State mixin ([#5][])
 * Push states in the browser history when changing day or track ([#6][])
 * Encode talk titles ([#7][])
+* Use the PureRenderMixin to improve performance ([#8][])
 * Code updated to use React 0.13.x
 * Dependencies updated
 
@@ -15,6 +16,7 @@
 [#5]: https://github.com/AurelioDeRosa/ConfAgenda/issues/5
 [#6]: https://github.com/AurelioDeRosa/ConfAgenda/issues/6
 [#7]: https://github.com/AurelioDeRosa/ConfAgenda/issues/7
+[#8]: https://github.com/AurelioDeRosa/ConfAgenda/issues/8
 
 # 0.1.0 (2015-02-10) #